Initial changes for re-enabling smart card cross domain logon attempts using marshaled credentials from certificate
